==> 2-isolate_neurons.Rmd <==
```{r}
library(Seurat)
library(SeuratObject)
library(scrubletR)
Sys.setenv(RETICULATE_PYTHON = "/Users/ananyashah/opt/anaconda3/envs/scrublet/bin/python")
library(reticulate)
use_condaenv("scrublet")

folder.names <-c("E10.5_WT", "E12.5_WT1", "E12.5_WT2", "E14.5_WT1", "E14.5_WT2")
data.path <- '/Volumes/backup-harddrive/Ananya/'
objs <- c()
for (f in folder.names) {
  path <- paste0(data.path, f)
  data <- Read10X(path)
  obj.temp <- CreateSeuratObject(counts = data, project = f)
  obj.temp <- scrublet_R(seurat_obj = obj.temp, python_home = "/Users/ananyashah/opt/anaconda3/envs/scrublet/bin/python" )  
  obj.temp <- NormalizeData(obj.temp)
  obj.temp <- FindVariableFeatures(obj.temp)
  obj.temp <- ScaleData(obj.temp)
  obj.temp <- RunPCA(obj.temp)
  obj.temp <- FindNeighbors(obj.temp, dims = 1:30, reduction = "pca")
  obj.temp <- FindClusters(obj.temp, resolution = 2, cluster.name = "unintegrated_clusters")
  obj.temp <- RunUMAP(obj.temp, dims = 1:30, reduction = "pca", reduction.name = "umap.unintegrated")
  objs <- c(objs, obj.temp)
}
```

```{r}
saveRDS(objs, '/Volumes/backup-harddrive/2-isolate_neurons_objs.RDS')
```

```{r}
objs = readRDS('/Volumes/backup-harddrive/2-isolate_neurons_objs.RDS')
```


```{r, fig.width=10, fig.height=12}

for (o in objs){
p2 <- DimPlot(o, reduction = "umap.unintegrated", group.by = c("unintegrated_clusters"),label=TRUE)
p3 <-DimPlot(o, reduction = "umap.unintegrated", group.by = c("predicted_doublets"),label=TRUE)
p4<- VlnPlot(o, "nFeature_RNA", group.by = "unintegrated_clusters")
p5<-VlnPlot(o, "nCount_RNA", group.by = "unintegrated_clusters")
print(p2)
print(p3)
print(p4)
print(p5)

neuron_markers <-c('Tubb3', 'Elavl4', 'Ret', 'Phox2b', 'Gap43', 'Elavl2', 'Slc18a3', 'Chat', 'Penk', 'Calb1', 'Calca', 'Calcb', 'Htr3b', 'Nmu', 'Chrnb4', 'Eml5', 'Smpd3', 'Tagln3', 'Snap25', 'Gpr22', 'Gdap1l1', 'Stmn3', 'Chrna3', 'Scg3', 'Syt4', 'Ncan', 'Crmp1', 'Adcyap1r1', 'Elavl3', 'Dlg2', 'Cacna2d', 'Uchl1' )
glia_markers <-c('Erbb3', 'Sox10', 'Fabp7', 'Plp1', 'Ezh2', 'Gas7', 'Nid1', 'Qk', 'Sparc', 'Mest', 'Nfia', 'Wwtr1', 'Gpm6b', 'Rasa3', 'Flrt1', 'Itpripl1', 'Itga4', 'Lama4', 'Postn', 'Ptprz1', 'Pdpn', 'Col18a1', 'Nrcam', 'S100b' )

nonneuronal_markers <-c('Epcam', 'Cdh1', 'Cdx2', 'Krt19', 'Ihh', 'Chga', 'Scgn' ,'Chgb', 'Neurod1', 'Lgr5', 'Axin2', 'Mki67', 'Col1a1', 'Col14a1', 'Cdk1',  'Dek', 'Aurkb', 'Hmmr', 'Myh11', 'Lmod1', 'Cnn1', 'Acta2' )
library(ggplot2)
p6 <- DotPlot(o, c(neuron_markers, glia_markers, nonneuronal_markers), scale.max = 75 , scale = F) + coord_flip() 
print(p6)
}
```

```{r}
objs_neurons <- c()
neuron_idx = list(c(7, 30, 31, 35), c(13, 35, 37, 41), c(6, 15, 20, 31, 32), c(13, 18, 23, 24, 25, 35), c(6, 19, 30, 35))
for (o in 1:5){
  Idents(objs[[o]]) = 'unintegrated_clusters'
  objs[[o]]$class_label = 'other'
  objs[[o]]$class_label[objs[[o]]$unintegrated_clusters %in% neuron_idx[[o]]] = 'neurons'
  obj_neurons = subset(objs[[o]], idents= neuron_idx[[o]])
  objs_neurons <- c(objs_neurons, obj_neurons)
}
obj = merge(objs_neurons[[1]], objs_neurons[2:5])
```

```{r}
obj$percent.mt <- PercentageFeatureSet(obj, pattern = "mt-")
obj <- JoinLayers(obj)
cell_mask <- Reduce(intersect,list(WhichCells(obj, expression = nFeature_RNA > 700),
                                   WhichCells(obj, expression = nFeature_RNA < 10000),
                                   WhichCells(obj, expression = nCount_RNA < 50000),
                                   WhichCells(obj, expression = percent.mt <12.5)))
gene_mask <- rownames(obj)[Matrix::rowSums(obj@assays$RNA@layers$counts > 0) > 8]
obj <- subset(obj, features = gene_mask, cells = cell_mask)
```

```{r}
obj <- split(obj, f = obj$orig.ident)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)
obj <- FindNeighbors(obj, dims = 1:30, reduction = "pca")
obj <- FindClusters(obj, resolution = 2, cluster.name = "unintegrated_clusters")
obj <- RunUMAP(obj, dims = 1:30, reduction = "pca", reduction.name = "umap.unintegrated")
```

```{r}
DimPlot(obj, reduction = "umap.unintegrated", group.by = c("orig.ident"),shuffle=TRUE)
DimPlot(obj, reduction = "umap.unintegrated", group.by = c("unintegrated_clusters"),label=TRUE)
DimPlot(obj, reduction = "umap.unintegrated", group.by = c("predicted_doublets"),label=TRUE)
```

```{r}
obj <- IntegrateLayers(
  object = obj, method = CCAIntegration,
  k.weight = 50, 
  orig.reduction = "pca", new.reduction = "integrated.cca",
  verbose = FALSE
)
obj <- FindNeighbors(obj, reduction = "integrated.cca", dims = 1:30)
obj <- FindClusters(obj, resolution = 1, cluster.name = "cca_clusters")
obj <- RunUMAP(obj, reduction = "integrated.cca", dims = 1:30, reduction.name = "umap.cca")
```

```{r}
DimPlot(obj, reduction = "umap.cca", group.by = "orig.ident", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "predicted_doublets", label = TRUE)
FeaturePlot(obj, "nCount_RNA", reduction = "umap.cca")
```

```{r}
obj$class <- NA
obj$class[obj$cca_clusters %in% c(0, 1, 3, 5, 6, 8, 10, 12, 13)] = 'known'
obj$class[obj$cca_clusters %in% c(2, 4, 7, 9, 11, 14, 15, 16)] = 'unknown'
```

```{r}
Idents(obj) = 'class'
obj = subset(obj, idents='known')
```

```{r}
#obj <- split(obj, f = obj$orig.ident)
obj <- NormalizeData(obj)
obj <- FindVariableFeatures(obj)
obj <- ScaleData(obj)
obj <- RunPCA(obj)
obj <- FindNeighbors(obj, dims = 1:30, reduction = "pca")
obj <- FindClusters(obj, resolution = 2, cluster.name = "unintegrated_clusters")
obj <- RunUMAP(obj, dims = 1:30, reduction = "pca", reduction.name = "umap.unintegrated")

obj <- IntegrateLayers(
  object = obj, method = CCAIntegration,
  k.weight = 50, 
  orig.reduction = "pca", new.reduction = "integrated.cca",
  verbose = FALSE
)
obj <- FindNeighbors(obj, reduction = "integrated.cca", dims = 1:30)
obj <- FindClusters(obj, resolution = 1.5, cluster.name = "cca_clusters")
obj <- RunUMAP(obj, reduction = "integrated.cca", dims = 1:30, reduction.name = "umap.cca")
```
```{r}
obj = JoinLayers(obj)
convert_human_to_mouse <- function(gene_names) {
  sapply(gene_names, function(x) { paste0(toupper(substr(x, 1, 1)), tolower(substr(x, 2, nchar(x)))) }) }
s.genes <- convert_human_to_mouse(cc.genes$s.genes)
g2m.genes <- convert_human_to_mouse(cc.genes$g2m.genes)
obj <- CellCycleScoring(obj, s.features = s.genes, g2m.features = g2m.genes, set.ident = FALSE)
```

```{r}
DimPlot(obj, reduction = "umap.cca", group.by = "orig.ident", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "Phase", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "predicted_doublets", label = TRUE)
FeaturePlot(obj, "nCount_RNA", reduction = "umap.cca")

# look at class markers
FeaturePlot(obj, "Mef2c", reduction = "umap.cca")
FeaturePlot(obj, "Btg2", reduction = "umap.cca")
FeaturePlot(obj, "Mllt11", reduction = "umap.cca")
```

```{r}
obj$class <- NA
obj$class[obj$cca_clusters %in% c(0, 1, 3, 4, 5, 9, 10, 11, 14)] = 'progenitor'
obj$class[obj$cca_clusters %in% c(2,6)] = 'neuroblast'
obj$class[obj$cca_clusters %in% c(7,8,12, 13)] = 'neuron'
```

```{r}
DimPlot(obj, reduction = "umap.cca", group.by = "orig.ident", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal()
DimPlot(obj, reduction = "umap.cca", group.by = "cca_clusters", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal()
DimPlot(obj, reduction = "umap.cca", group.by = "predicted_doublets", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal()
DimPlot(obj, reduction = "umap.cca", group.by = "class", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal()
FeaturePlot(obj, "nCount_RNA", reduction = "umap.cca") + xlim(-10,7) + ylim(-8, 9) + coord_equal()

p = DimPlot(obj, reduction = "umap.cca", group.by = "class", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal() 
ggsave('/Volumes/backup-harddrive/figures/figure 1/neuron_class_umap.svg', plot = p )
```
```{r}
saveRDS(obj, '/Volumes/backup-harddrive/2-isolate_neurons_output.RDS')
```
```{r}
obj = readRDS('/Volumes/backup-harddrive/2-isolate_neurons_output.RDS')
```
