==> 4-timepoint_types.Rmd <==

```{r}
obj  = readRDS("/Volumes/backup-harddrive/2-isolate_neurons_output.RDS")
obj = JoinLayers(obj)
obj@assays$RNA@layers$data <- NULL
obj@assays$RNA@layers$scale.data <- NULL
```

```{r}
obj$timepoint <- NA
obj$timepoint[obj$orig.ident %in% c('E10.5_WT')] = '10.5'
obj$timepoint[obj$orig.ident %in% c('E12.5_WT2', 'E12.5_WT1')] = '12.5'
obj$timepoint[obj$orig.ident %in% c('E14.5_WT1', 'E14.5_WT2')] = '14.5'
```

```{r}
counts.tc <- GetAssayData(obj, assay = "RNA", layer = "counts")
obj[["RNA"]] <- CreateAssay5Object(counts = counts.tc)

obj = split(obj, f = obj$orig.ident)
obj = NormalizeData(obj)
obj = FindVariableFeatures(obj)
obj = ScaleData(obj)
obj = RunPCA(obj)
obj <- IntegrateLayers(
  object = obj, method = CCAIntegration,
  k.weight = 50, 
  orig.reduction = "pca", new.reduction = "integrated.cca",
  verbose = FALSE
)

obj <- FindNeighbors(obj, reduction = "integrated.cca", dims = 1:30)
obj <- FindClusters(obj, resolution = 1.5, cluster.name = "cca_clusters")
obj <- RunUMAP(obj, reduction = "integrated.cca", dims = 1:30, reduction.name = "umap.cca")
```

```{r}
type_genes = c('Mef2c', 'Btg2',"Etv1", "Vip", "Calb1" , "Bnc2", "Ndufa4l2" , "Ano2", "Sema5a", "Penk", "Mgat4c", "Gal","Nos1", "Ntng1", "Slc18a2" , "Gda")
for(g in type_genes){
  p = FeaturePlot(obj, g, reduction = "umap.cca")
  print(p)
}
Idents(obj) = 'cca_clusters'
DotPlot(obj, type_genes, scale = F) + coord_flip()
```

```{r}
DimPlot(obj, reduction = "umap.cca", group.by = "class", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "orig.ident", label = TRUE)
DimPlot(obj, reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
```


```{r}
obj$type = NA
obj$type[obj$cca_clusters %in% c(8)] = 'iMN_8/9'
obj$type[obj$cca_clusters %in% c(13)] = 'iMN_12' 
obj$type[obj$cca_clusters %in% c(7)] = 'eMN_1' 
obj$type[obj$cca_clusters %in% c(12)] = 'eMN_7' 
obj$type[obj$cca_clusters %in% c(0, 1, 3, 4, 5, 9, 10, 11, 14)] = 'progenitor' 
obj$type[obj$cca_clusters %in% c(2,6)] = 'neuroblast' 
```

```{r}
p = DimPlot(obj, reduction = "umap.cca", group.by = "type", label = TRUE) + xlim(-10,7) + ylim(-8, 9) + coord_equal()
ggsave('/Volumes/backup-harddrive/figures/figure 2/neuron_type_umap.svg', plot = p )
print(p)
```

```{r}
cluster.sample <- table(obj$type, obj$timepoint) %>%
    as.data.frame.matrix() %>%
    rownames_to_column(var = "sample")
  cluster.sample[-1] <- lapply(cluster.sample[-1], function(x) x/sum(x))
  cluster.sample <- cluster.sample %>%
    pivot_longer(
      cols = -c("sample"),
      names_to = "cluster",
      values_to = "count"
    )
  cluster.sample$cluster <- factor(cluster.sample$cluster, levels = unique(cluster.sample$cluster))
  cluster.sample$sample <- factor(cluster.sample$sample, levels = c('eMN_7', 'eMN_1', 'iMN_12', 'iMN_8/9', 'neuroblast','progenitor'))
p = ggplot(cluster.sample, aes(x=cluster, y=count, fill=sample)) +
    geom_bar(stat="identity") +
    theme_minimal()  
print(p)
#gsave(paste0('/Volumes/backup-harddrive/figures/figure 2/neuron_type_bar.svg'), plot = p)

```
```{r}
obj.neuron = subset(obj, class=='neuron')
cluster.sample <- table(obj.neuron$type, obj.neuron$timepoint) %>%
    as.data.frame.matrix() %>%
    rownames_to_column(var = "sample")
  cluster.sample[-1] <- lapply(cluster.sample[-1], function(x) x/sum(x))
  cluster.sample <- cluster.sample %>%
    pivot_longer(
      cols = -c("sample"),
      names_to = "cluster",
      values_to = "count"
    )
  cluster.sample$cluster <- factor(cluster.sample$cluster, levels = unique(cluster.sample$cluster))
  cluster.sample$sample <- factor(cluster.sample$sample, levels = c('eMN_7', 'eMN_1', 'iMN_12', 'iMN_8/9'))
p = ggplot(cluster.sample, aes(x=cluster, y=count, fill=sample)) +
    geom_bar(stat="identity") +
    theme_minimal()  
print(p)
#gsave(paste0('/Volumes/backup-harddrive/figures/figure 2/only_neuron_type_bar.svg'), plot = p)

```

```{r}
timepoints = c('10.5', '12.5', '14.5')
obj.tc <- setNames(vector("list", length(timepoints)), timepoints)
for (t in timepoints){
  Idents(obj) = 'timepoint'
  obj.tc[[t]] = subset(obj, idents=t)
  counts = GetAssayData(obj.tc[[t]], assay = "RNA", layer = "counts")
  obj.tc[[t]][["RNA"]] = CreateAssay5Object(counts = counts)
  if (length(unique(obj.tc[[t]]$orig.ident)) > 1) {
    obj.tc[[t]] = split(obj.tc[[t]], f = obj.tc[[t]]$orig.ident)
  }
  obj.tc[[t]] = NormalizeData(obj.tc[[t]])
  obj.tc[[t]] = FindVariableFeatures(obj.tc[[t]])
  obj.tc[[t]] = ScaleData(obj.tc[[t]])
  if (length(unique(obj.tc[[t]]$orig.ident)) > 1) {
    obj.tc[[t]] <- IntegrateLayers(
      object = obj.tc[[t]], method = CCAIntegration,
      k.weight = 50,
      orig.reduction = "pca", new.reduction = "integrated.cca",
      verbose = FALSE
         )
  }
  else {
     obj.tc[[t]] = RunPCA(obj.tc[[t]], reduction.name = "integrated.cca")
      }
  obj.tc[[t]] <- FindNeighbors(obj.tc[[t]], reduction = "integrated.cca", dims = 1:30)
  obj.tc[[t]] <- FindClusters(obj.tc[[t]], resolution = 1.5, cluster.name = "cca_clusters")
  obj.tc[[t]] <- RunUMAP(obj.tc[[t]], reduction = "integrated.cca", dims = 1:30, reduction.name = "umap.cca")
}
```

```{r}
type_genes = c('Mef2c', 'Btg2',"Etv1", "Vip", "Calb1" , "Bnc2", "Ndufa4l2" , "Ano2", "Sema5a", "Penk", "Mgat4c", "Gal","Nos1", "Ntng1", "Slc18a2" , "Gda")
DimPlot(obj.tc[['10.5']], reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
DimPlot(obj.tc[['10.5']], reduction = "umap.cca", group.by = "class", label = TRUE)
for(g in type_genes){
  p = FeaturePlot(obj.tc[['10.5']], g, reduction = "umap.cca")
  print(p)
}
Idents(obj.tc[['10.5']]) = 'cca_clusters'
DotPlot(obj.tc[['10.5']], type_genes, scale = F) + coord_flip()
```

```{r}
obj.tc[['10.5']]$type = NA
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(15, 16)] = 'iMN_8'
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(5)] = 'iMN_9'
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(17)] = 'iMN_12' 
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(10, 11)] = 'eMN_1' 
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(14)] = 'eMN_7' 
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(0, 2, 3, 4, 6, 7, 8, 9, 12, 13)] = 'progenitor' 
obj.tc[['10.5']]$type[obj.tc[['10.5']]$cca_clusters %in% c(1)] = 'neuroblast' 
```


```{r}
type_genes = c('Mef2c', 'Btg2',"Etv1", "Vip", "Calb1" , "Bnc2", "Ndufa4l2" , "Ano2", "Sema5a", "Penk", "Mgat4c", "Gal","Nos1", "Ntng1", "Slc18a2" , "Gda")
DimPlot(obj.tc[['12.5']], reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
DimPlot(obj.tc[['12.5']], reduction = "umap.cca", group.by = "class", label = TRUE)
for(g in type_genes){
  p = FeaturePlot(obj.tc[['12.5']], g, reduction = "umap.cca")
  print(p)
}
Idents(obj.tc[['12.5']]) = 'cca_clusters'
DotPlot(obj.tc[['12.5']], type_genes, scale = F)
```

```{r}
obj.tc[['12.5']]$type = NA
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(15, 16)] = 'iMN_8'
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(5)] = 'iMN_9'
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(17)] = 'iMN_12' 
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(10, 11)] = 'eMN_1' 
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(14)] = 'eMN_7' 
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(0, 2, 3, 4, 6, 7, 8, 9, 12, 13)] = 'progenitor' 
obj.tc[['12.5']]$type[obj.tc[['12.5']]$cca_clusters %in% c(1)] = 'neuroblast' 
```

```{r}
type_genes = c('Mef2c', 'Btg2',"Etv1", "Vip", "Calb1" , "Bnc2", "Ndufa4l2" , "Ano2", "Sema5a", "Penk", "Mgat4c", "Gal","Nos1", "Ntng1", "Slc18a2" , "Gda")
DimPlot(obj.tc[['14.5']], reduction = "umap.cca", group.by = "cca_clusters", label = TRUE)
for(g in type_genes){
  p = FeaturePlot(obj.tc[['14.5']], g, reduction = "umap.cca")
  print(p)
}
Idents(obj.tc[['14.5']]) = 'cca_clusters'
DotPlot(obj.tc[['14.5']], type_genes, scale = F)
```

```{r}
obj.tc[['14.5']]$type = NA
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(15, 16)] = 'iMN_8'
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(5)] = 'iMN_9'
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(17)] = 'iMN_12' 
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(10, 11)] = 'eMN_1' 
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(14)] = 'eMN_7' 
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(0, 2, 3, 4, 6, 7, 8, 9, 12, 13)] = 'progenitor' 
obj.tc[['14.5']]$type[obj.tc[['14.5']]$cca_clusters %in% c(1)] = 'neuroblast' 
```

```{r}
xlimits = list('10.5'=c(-8,6), '12.5' = c(-6,8), '14.5' = c(-9, 13))
ylimits = list('10.5' = c(-6,8), '12.5' = c(-6,8), '14.5' = c(-11, 11))
for (g in c('Gal', 'Ntng1', 'Ndufa4l2', 'Sema5a')){
for (t in timepoints){
  p = FeaturePlot(obj.tc[[t]], g, reduction = "umap.cca", max.cutoff = 2) + NoLegend() + xlim(xlimits[[t]][1], xlimits[[t]][2]) +   ylim(ylimits[[t]][1], xlimits[[t]][2]) + coord_equal()
  print(p)
  ggsave(paste0('/Volumes/backup-harddrive/figures/figure 2/', t, '_', g ,'_', 'neuron_ feature_umap.svg'), plot = p )
}}
```
